Removed javascript validate validateForm()
This validation was used when we used textbox as the method user input. In that case if the "userid" entered was empty, the validateForm() would throw a pop to warn the user.
Now we use a dropdown-list. The "userid" will now have a default non null value. Hence, validateForm() is not needed.